Update context in index.js.
The app started printing 'Generate a' in front of the poem title. It was not doing this originally, so I added context to try to stop that as well. Originally it was doing as directed, printing a title in bold then 4-8 lines about the input topic. Maybe they've changed their API?